Split some more objects into their own files. added overloads to item.
